# skyline-rush-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Skyline Rush API
  version: 1.0.0
  description: >
    Official REST API specification for Skyline Rush, an original iOS/iPadOS
    3-lane endless runner. All mutating economy and run endpoints require an
    Idempotency-Key header.
servers:
  - url: https://api.skylinerush.game
    description: Production Gateway
  - url: http://localhost:3000
    description: Local Dev Gateway
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access JWT issued by Profile & Auth service (15-minute expiry).
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: INSUFFICIENT_BALANCE
            message:
              type: string
              example: Not enough Cores for this Redeploy.
            details:
              type: object
              additionalProperties: true
...
    AuthGuestRequest:
      type: object
      required:
        - guest_device_id
        - age_bucket
      properties:
        guest_device_id:
          type: string
          format: uuid
        age_bucket:
          $ref: '#/components/schemas/AgeBucket'
    AuthAppleRequest:
      type: object
      required:
        - identity_token
      properties:
        identity_token:
          type: string
        guest_device_id:
          type: string
          format: uuid
          nullable: true
    AuthResponse:
      type: object
      required:
        - player_id
        - access_token
        - refresh_token
      properties:
        player_id:
          type: string
          format: uuid
        access_token:
          type: string
        refresh_token:
          type: string
        age_bucket:
          $ref: '#/components/schemas/AgeBucket'
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
...
paths:
  /v1/auth/guest:
    post:
      summary: Create or resume guest session
      operationId: authGuest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthGuestRequest'
      responses:
        '200':
          description: Guest session authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/apple:
    post:
      summary: Link or resume via Sign in with Apple
      operationId: authApple
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthAppleRequest'
      responses:
        '200':
          description: Apple account authenticated / merged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/refresh:
    post:
      summary: Exchange refresh token for new access token
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/auth/parental-gate/verify:
    post:
      summary: Verify parental gate challenge answer and receive 5-minute gate token
      operationId: verifyParentalGate
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentalGateVerifyRequest'
      responses:
        '200':
          description: Parental gate verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentalGateVerifyResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Incorrect answer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/profile:
    get:
      summary: Retrieve current authenticated player profile
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Player profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/runs:
    post:
      summary: Submit a completed run
      operationId: submitRun
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunSubmitRequest'
      responses:
        '201':
          description: Run accepted and scored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunSubmitResponse'
        '400':
          description: Validation error / Implausible run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/runs/{run_id}/redeploy:
    post:
      summary: Revive/redeploy runner during a run
      operationId: redeployRun
      security:
        - BearerAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeployRequest'
      responses:
        '200':
          description: Redeploy granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedeployResponse'
        '402':
          description: Insufficient Core balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Run already ended or conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/runs/redeploy:
    post:
      summary: Revive/redeploy runner during a run (run_id in body)
      operationId: redeployRunInBody
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeployRequest'
      responses:
        '200':
          description: Redeploy granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedeployResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Insufficient Core balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Run not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Run already ended or conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/economy/balance:
    get:
      summary: Get player current balances
      operationId: getBalance
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Balances retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EconomyBalanceResponse'
  /v1/economy/ledger:
    get:
      summary: Get audit ledger history
      operationId: getLedger
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Ledger entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerResponse'
  /v1/contracts/active:
    get:
      summary: List active daily and weekly contracts
      operationId: getActiveContracts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Active contracts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveContractsResponse'
  /v1/contracts/{contract_id}/claim:
    post:
      summary: Claim contract reward
      operationId: claimContract
      security:
        - BearerAuth: []
      parameters:
        - name: contract_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '200':
          description: Contract claimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimContractResponse'
        '409':
          description: Contract not yet completed or already claimed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/supply-drops/tables/{table_id}:
    get:
      summary: Disclose exact odds table for Supply Drop
      operationId: getSupplyDropTable
      parameters:
        - name: table_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Supply drop odds table
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplyDropTableResponse'
  /v1/supply-drops/open:
    post:
      summary: Open a Supply Drop (earned or purchased)
      operationId: openSupplyDrop
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyDropOpenRequest'
      responses:
        '200':
          description: Supply drop outcome resolved and granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplyDropOpenResponse'
  /v1/roster:
    get:
      summary: List player runners and boards
      operationId: getRoster
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Roster items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RosterResponse'
  /v1/roster/equip:
    post:
      summary: Equip a runner or board
      operationId: equipRosterItem
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RosterEquipRequest'
      responses:
        '200':
          description: Equipped item
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
  /v1/roster/unlock:
    post:
      summary: Unlock a runner or board with currency
      operationId: unlockRosterItem
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RosterUnlockRequest'
      responses:
        '200':
          description: Item unlocked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  balance:
                    $ref: '#/components/schemas/EconomyBalanceResponse'
        '402':
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/leaderboard:
    get:
      summary: Query District leaderboard
      operationId: getLeaderboard
      security:
        - BearerAuth: []
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [global, friends]
            default: global
        - name: district_id
          in: query
          required: false
          schema:
            type: string
            default: neo-marina
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Leaderboard ranking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
...